Adds file spillage code.
If an output gets too far behind, data will be written to a temporary file
and read back again later instead of keeping it all in memory.  In the
case where the input is a regular file, the temporary file is not needed as
the data can simply be re-read from the input.

Various bug fixes, including:

  Reversing the order of closing files to avoid the indexes into the regular
and pipes arrays becoming incorrect.

  Ensuring pipes are closed even when they don't need to be polled.  Also
means outputs don't need to be added to the poll list if they have finished
and there is no more data to read.

  Ensuring all of the data has been written after the read has finished.

  Properly releasing any left-over data after output files have closed.