Reworks spillage and adds an option to defer spilling.
Spillage gets its own struct to store the total allocated, the current
spill file and a linked list of non-regular outputs which is kept sorted
by the number of bytes written.

Adds -w option, which is the number of seconds to wait for a blocked
output before spilling data.

Adds code to keep track of which non-regular output has sent the fewest
bytes, and when they were written.  If the memory limit is hit, but the
output last sent data within the -w limit, spillage is deferred for a while.
This hopefully prevents data from being spilled when one output is being
read slightly more slowly than the others, at the expense of making all
the outputs run at the speed of the slowest one.