Fix bug with removal of items from pipes array in do_copy.
The loop that closed poll-able files in do_copy also removed the corresponding
entries from the pipes array.  To do this successfully, it was necessary to
remove the entries in the correct order to ensure that later look-ups didn't
try to access an entry that had moved.  For that to be the case it was
necessary for the values in the closing_pipes array to be in ascending order.
Unfortunately, in certain rare conditions this was not true, causing either
premature file closure or assertion failures.

This fixes the problem by flagging the entries that need to be removed, and
then removing them in bulk later.  Both the pipes and regular files versions
of the closing loop are updated to use the new method, and the part that
removes the array elements is in a new function cull_index_array().  Some
extra assertions have also been put in to ensure the new method is working
as intended.