Multi-threaded version.
Add options -p / -P to turn multi-threading on / off.
Multi-threaded mode starts a new thread for each writer.  The reader runs in
the main thread.  Having the writers in separate threads means there is
no need to wait for all the opens to complete before starting.  It also
makes the output appear slightly faster (at least on multi-cpu machines)
as a writer doesn't have to wait for other write jobs to finish before it
starts.

Adds lots of mutexes and conditionals plus some new functions to run the
multi-threaded copy.  Code that requires pthreads is wrapped in #ifdef
HAVE_PTHREADS guards, plus there are some macros for locking and unlocking
mutexes that expand to the real code if HAVE_PTHREADS is defined and
empty blocks if it isn't.